requirements.txt
docxtpl==0.19.0
python-docx==1.1.2
Pillow==10.4.0
num2words==0.5.13
Babel==2.14.0